pyproject.toml
[project]
name = "kalorietracker"
version = "0.1.0"
description = "A Nigerian-focused calorie tracking and meal planning app"
authors = [
    {name = "Aniekan Umanah"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "scalar-fastapi (>=1.0.3,<2.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "psycopg[binary] (>=3.2.7,<4.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
devtools = "^0.12.2"
ruff = "^0.12.0"
mypy = "^1.15.0"
types-passlib = "^1.7.7.20250602"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.0"
coverage = "^7.8.0"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_also = [
    'def __repr__',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
    '@(abc\.)?abstractmethod',
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "tests.fixtures.*"
disallow_untyped_calls = false

[tool.ruff]
target-version = "py311"
exclude = [
    "migrations",
    "tests",
]

[tool.ruff.lint]
select = [
    "E", 
    "F", 
    "I", 
    "B", 
    "FAST", 
    "UP"
]
ignore = [
  "B008",  # flake8-bugbear false positives with FastAPI Depends(), 
  "B904",  # Allow raising exceptions without from e, for HTTPException, 
  "FAST002",  # Allow FastAPI Depends() in function signatures without 'Annotated',
]